Separate API from impl (WIP)
