+++ api/src/main/kotlin/io/github/pak3nuh/messaging/outbox/Outbox.kt
@@ -1,0 +1,6 @@
+package io.github.pak3nuh.messaging.outbox
+
+interface Outbox : AutoCloseable {
+    fun submit(entry: Entry)
+    fun start()
+}
