+++ integration-test/src/test/kotlin/io/github/pak3nuh/messaging/outbox/storage/MySqlEntryStorageTest.kt
@@ -16,12 +16,14 @@ class MySqlEntryStorageTest {
         val host = container.host
         val port = container.firstMappedPort
         val entryStorage = SqlEntryStorage("jdbc:mysql://$host:$port/database", "mysql", "mysql")
-        entryStorage.persist(Entry(
+        entryStorage.persist(
+            Entry(
             byteArrayOf(1),
             byteArrayOf(2),
             "should-store-entry",
             mapOf(Pair("version","123"))
-        ))
+        )
+        )
 
         val batch = entryStorage.getBatch()
         assertEquals(1, batch.count())
