+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include("util")
 include("lock")
 include("test-util")
 include("integration-test")
+include("api")
